+++ flask_ades_wpst/sqlite_connector.py
@@ -80,7 +80,15 @@ def sqlite_get_proc(proc_id):
     sql_str = """SELECT * FROM processes
                  WHERE id = \"{}\"""".format(proc_id)
     cur.execute(sql_str)
-    return cur.fetchall()[0]
+    res = cur.fetchall()
+    num_matches = len(res)
+    if num_matches == 1:
+        res = res[0]
+    else:
+        assert(num_matches == 0, 
+               "Found more than one match for {}. This should never happen.".
+               format(proc_id))
+    return res
 
 @sqlite_db
 def sqlite_deploy_proc(proc_spec):
@@ -111,12 +119,13 @@ def sqlite_deploy_proc(proc_spec):
 @sqlite_db
 def sqlite_undeploy_proc(proc_id):
     proc_desc = sqlite_get_proc(proc_id)
-    conn = create_connection(db_name)
-    cur = conn.cursor()
-    sql_str = """DELETE FROM processes
-                 WHERE id = \"{}\"""".format(proc_id)
-    cur.execute(sql_str)
-    conn.commit()
+    if proc_desc:
+        conn = create_connection(db_name)
+        cur = conn.cursor()
+        sql_str = """DELETE FROM processes
+                     WHERE id = \"{}\"""".format(proc_id)
+        cur.execute(sql_str)
+        conn.commit()
     return proc_desc
 
 def sqlite_get_headers(cur, tname):
@@ -144,15 +153,24 @@ def sqlite_get_job(job_id):
     cur = conn.cursor()
     sql_str = """SELECT * FROM jobs
                  WHERE jobID = \"{}\"""".format(job_id)
-    job = cur.execute(sql_str).fetchall()[0]
-    col_headers = sqlite_get_headers(cur, "jobs")
-    job_dict = {}
-    for i, col in enumerate(col_headers):
-        # deserialize JSON data fields
-        if col in ("inputs", "backend_info"):
-            job_dict[col] = json.loads(job[i])
-        else:
-            job_dict[col] = job[i]
+    job_matches = cur.execute(sql_str).fetchall()
+    num_matches = len(job_matches)
+    if num_matches == 0:
+        job_dict = {}
+    elif num_matches == 1:
+        job = job_matches[0]
+        col_headers = sqlite_get_headers(cur, "jobs")
+        job_dict = {}
+        for i, col in enumerate(col_headers):
+            # deserialize JSON data fields
+            if col in ("inputs", "backend_info"):
+                job_dict[col] = json.loads(job[i])
+            else:
+                job_dict[col] = job[i]
+    else:
+        # This should never happen.
+        raise ValueError("Found more than one match for job ID {}".
+                         format(job_id))
     return job_dict
 
 @sqlite_db
@@ -183,4 +201,9 @@ def sqlite_update_job_status(job_id, status):
 
 @sqlite_db
 def sqlite_dismiss_job(job_id):
-    return sqlite_update_job_status(job_id, "dismissed")
+    job_dict = sqlite_get_job(job_id)
+    if job_dict:
+        resp = sqlite_update_job_status(job_id, "dismissed")
+    else:
+        resp = {}
+    return resp
